% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorers.R
\name{BNLearnScorer}
\alias{BNLearnScorer}
\title{BNLearnScorer}
\usage{
BNLearnScorer(node, parents, ...)
}
\arguments{
\item{node}{A child node of parents.}

\item{parents}{The parent nodes of node.}

\item{...}{The ellipsis is used to pass other parameters to the scorer.}
}
\description{
BNLearnScorer
}
\examples{
data <- bnlearn::learning.test
BNLearnScorer('A', c('B', 'C'), data = data)
BNLearnScorer('A', c(), data = data)
BNLearnScorer('A', vector(), data = data)
BNLearnScorer('A', NULL, data = data)
BNLearnScorer('A', c('B', 'C'), data = data, type = 'bde', iss = 100)
BNLearnScorer('A', c('B', 'C'), data = data, type = 'bde', iss = 1)

}
