% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{CalculateFeatureProbability}
\alias{CalculateFeatureProbability}
\title{Collect DAG feature probability.}
\usage{
CalculateFeatureProbability(collection, p_feature)
}
\arguments{
\item{collection}{A collection of unique objects. See CollectUniqueObjects.}

\item{p_feature}{A function that takes an adjacency matrix and collection object
and returns a numeric value equal to p(f|G). Therefore, it must be of the
form p_feature(dag).}
}
\value{
p_post_feature A numeric value representing the posterior probability
of the feature.
}
\description{
Calculate the feature (\eqn{f}) probability whereby
\eqn{p(f|D) = \sum_\mathcal{G \in G} p(G|D)p(f|G)}.
}
