% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_node_move.R
\name{CalculateNodeMoveNeighbourhood}
\alias{CalculateNodeMoveNeighbourhood}
\title{Calculate neighbourhood for node move.}
\usage{
CalculateNodeMoveNeighbourhood(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Calculate neighbourhood for node move.
}
