% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_split_join.R
\name{CalculateSplitCombinations}
\alias{CalculateSplitCombinations}
\title{Calculate number of split combinations.}
\usage{
CalculateSplitCombinations(partitioned_nodes)
}
\description{
Calculate number of split combinations.
}
