% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_split_join.R
\name{CalculateSplitJoinNeighbourhood}
\alias{CalculateSplitJoinNeighbourhood}
\title{Calculate neighbourhood for the split or join proposal.}
\usage{
CalculateSplitJoinNeighbourhood(partitioned_nodes)
}
\description{
Calculate neighbourhood for the split or join proposal.
}
