% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_stay.R
\name{CalculateStayStillNeighbourhood}
\alias{CalculateStayStillNeighbourhood}
\title{Calculate neighbourhood for staying still.}
\usage{
CalculateStayStillNeighbourhood(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}
}
\description{
Calculate neighbourhood for staying still.
}
