% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_swap_adjacent.R
\name{CalculateSwapAdjacentNodeNeighbourhood}
\alias{CalculateSwapAdjacentNodeNeighbourhood}
\title{Calculate neighbourhood for swapping nodes.}
\usage{
CalculateSwapAdjacentNodeNeighbourhood(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Calculate neighbourhood for swapping nodes.
}
