% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_swap_node.R
\name{CalculateSwapNodeNeighbourhood}
\alias{CalculateSwapNodeNeighbourhood}
\title{Calculate neighbourhood for swapping nodes.}
\usage{
CalculateSwapNodeNeighbourhood(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Calculate neighbourhood for swapping nodes.
}
