% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{CheckBlacklistObeyed}
\alias{CheckBlacklistObeyed}
\title{Check blacklist obeyed.}
\usage{
CheckBlacklistObeyed(partitioned_nodes, blacklist = NULL, nodes = NULL)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{blacklist}{A data.frame of (parent, child) pairs representing edges
that cannot be in the DAG.}

\item{nodes}{A vector of node names to check. Default is to check all
child nodes in the blacklist.}
}
\description{
If an edge between two nodes is blacklisted in Partition MCMC the adjacent
partition element cannot be the only direct node for it's blacklisted child.
}
