% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{CheckWhitelistObeyed}
\alias{CheckWhitelistObeyed}
\title{Check whitelist is obeyed.}
\usage{
CheckWhitelistObeyed(partitioned_nodes, whitelist = NULL, nodes = NULL)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{whitelist}{A data.frame of (parent, child) pairs representing edges
that must be in the DAG.}

\item{nodes}{A vector of node names to check. Default is to check all
child nodes in the whitelist.}
}
\description{
Check whitelist is obeyed.
}
