% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_package.R
\docType{package}
\name{cia-package}
\alias{cia}
\alias{cia-package}
\title{cia: Learn and Apply Directed Acyclic Graphs for Causal Inference}
\description{
Causal Inference Assistance (CIA) for performing causal inference within the structural causal modelling framework. Structure learning is performed using partition Markov chain Monte Carlo (Kuipers & Moffa, 2017) and several additional functions have been added to help with causal inference. Kuipers and Moffa (2017) \doi{10.1080/01621459.2015.1133426}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spaceodyssey.github.io/cia/}
  \item Report bugs at \url{https://github.com/SpaceOdyssey/cia/issues}
}

}
\author{
\strong{Maintainer}: Mathew Varidel \email{mathew.varidel@sydney.edu.au} (\href{https://orcid.org/0000-0002-1648-8317}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Victor An \email{victor.an@sydney.edu.au} [contributor]
}

}
\keyword{internal}
