% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{CollectDags}
\alias{CollectDags}
\title{Collect DAGs.}
\usage{
CollectDags(post_chain)
}
\arguments{
\item{post_chain}{A flattened chain that includes a DAG per sample.}
}
\value{
dag_collection A list with entries:
dag: A list of all unique DAGs within the sample.
log_score: A vector with the log_scores for each DAG.
log_evidence: A numeric value representing the evidence \eqn{log(\tilde{Z}) = \Sigma_s^S log(p(G_s)p(D | G_s))}.
log_norm_score: A vector of normalised log_scores for each DAG \eqn{G_s} using \eqn{p(G_s | D) = log(p(G_s)p(D | G_s)) - log(\tilde{Z})}.
}
\description{
Get the unique set of DAGs along with their log_score. It also estimates
the normalised log score assuming \eqn{\tilde{Z} = \Sigma_s^S p(G_s)p(D | G_s)}
where \eqn{\{G_1, G_2, G_3, ..., G_S\}} is the set of unique DAGs in the
the chain. This assumes that you have captured the most probable graphs, such
that \eqn{\tilde{Z}} is approximately equal to the true evidence
\eqn{Z = \Sigma_{G \in \mathcal{G}} p(G)p(D | G)}
where you sum across all possible graphs \eqn{\mathcal{G}}. This
also makes the assumption that the scoring method used is proportional to
the posterior probability; \eqn{\text{score}(G, D) \propto p(G)p(G | D)}.
}
