% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmc-package.R
\docType{package}
\name{dagmc-package}
\alias{dagmc}
\alias{dagmc-package}
\title{dagmc: MCMC sampler for Directed Acyclic Graphs.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mathew Varidel \email{mathew.varidel@sydney.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
