% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_to_cpdag.R
\name{DAGtoCPDAG}
\alias{DAGtoCPDAG}
\title{Convert DAG to CPDAG.}
\usage{
DAGtoCPDAG(x)
}
\arguments{
\item{x}{A matrix, dagmc_chain, or dagmc_chains object. When it is a chain(s)
object the state must be an adjacency matrix.}
}
\value{
x Returns same object type converted to a CPDAG.
}
\description{
Convert DAG to CPDAG.
}
