% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{FindChangedNodes}
\alias{FindChangedNodes}
\title{Find nodes with changed parent combinations between different labelled
partitions.}
\usage{
FindChangedNodes(old_partitioned_nodes, new_partitioned_nodes)
}
\arguments{
\item{old_partitioned_nodes}{A labelled partition.}

\item{new_partitioned_nodes}{A labelled partition.}
}
\value{
changed_nodes A vector of changed nodes.
}
\description{
Find nodes with changed parent combinations between different labelled
partitions.
}
\examples{
changed_nodes <- FindChangedNodes(partitioned_nodes, new_partitioned_nodes)

}
