% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dags.R
\name{GetEmptyDAG}
\alias{GetEmptyDAG}
\title{Get an empty DAG given a set of nodes.}
\usage{
GetEmptyDAG(nodes)
}
\arguments{
\item{nodes}{A vector of node names.}
}
\value{
An adjacency matrix with elements designated as (parent, child).
}
\description{
Get an empty DAG given a set of nodes.
}
\examples{
GetEmptyDAG(LETTERS[1:3])

}
