% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{GetIncrementalScoringEdges}
\alias{GetIncrementalScoringEdges}
\title{Get incremental edges}
\usage{
GetIncrementalScoringEdges(scorer, cutoff = 0)
}
\arguments{
\item{scorer}{A scorer object.}

\item{cutoff}{A score cutoff.}
}
\value{
A Boolean matrix of (parent, child) pairs for blacklisting.
}
\description{
Get edges that incrementally improve the score over an empty DAG.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- DAGtoPartition(dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )
  
blacklist <- GetIncrementalScoringEdges(scorer, cutoff = -10.0)

blacklist_scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )

results <- SampleChains(10, partitioned_nodes, PartitionMCMC(), blacklist_scorer)

}
