% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetLowerPartitionedNodes}
\alias{GetLowerPartitionedNodes}
\title{Get potential parents of a node.}
\usage{
GetLowerPartitionedNodes(node, partitioned_nodes)
}
\arguments{
\item{node}{A node name.}

\item{partitioned_nodes}{A labelled partition.}
}
\value{
potential_parents A vector of potential parent names.
}
\description{
Get potential parents of a node.
}
