% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{GetLowestScoringEdges}
\alias{GetLowestScoringEdges}
\title{Preprocessing for blacklisting.
Get the lowest pairwise scoring edges.}
\usage{
GetLowestScoringEdges(scorer, c)
}
\arguments{
\item{scorer}{A scorer object.}

\item{c}{An integer representing the number of edges to retain.}
}
\description{
Get the lowest c pairwise scoring edges represented as a blacklist matrix.
}
