% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{GetMAP}
\alias{GetMAP}
\title{Get MAP DAG. TODO: I'm confused here too. Why should this specifically run on
the chain collection? Why not the chains or list of chains? What if I want to
check that the MAP is the same across chains? Why wouldn't this give back
the map per chain in that case? What if we have more complex 'states' that
aren't described by a single DAG (e.g., mixtures of DAGs).}
\usage{
GetMAP(x)
}
\arguments{
\item{x}{A collection of unique objects. See CollectUniqueObjects.}
}
\value{
dag A list with the adjacency matrix for the map and it's posterior
probability. It is possible for it to return multiple DAGs. The list has
elements;
\itemize{
\item dag: List of MAP DAGs.
\item log_p: Vector with the log posterior probability for each DAG.
}
}
\description{
Get the maximum a posteriori DAG.
}
