% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{GetMAPDAG}
\alias{GetMAPDAG}
\title{Get MAP DAG.}
\usage{
GetMAPDAG(collection)
}
\arguments{
\item{collection}{A collection of unique objects. See CollectUniqueObjects.}
}
\value{
dag A list with the adjacency matrix for the map and it's posterior
probability. It is possible for it to return multiple DAGs. The list has
elements;
\itemize{
\item dag: List of MAP DAGs.
\item log_p: Vector with the log posterior probability for each DAG.
}
}
\description{
Get the maximum a posteriori DAG.
}
