% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetNodeCombinations}
\alias{GetNodeCombinations}
\title{Get all combinations of nodes.}
\usage{
GetNodeCombinations(nodes)
}
\arguments{
\item{nodes}{Vector of node names.}
}
\value{
List of node combinations.
}
\description{
Get all combinations of nodes.
}
