% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetNodePartition}
\alias{GetNodePartition}
\title{Get a node's partition element number.}
\usage{
GetNodePartition(partitioned_nodes, node)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{node}{Node name.}
}
\value{
Node's partition element number.
}
\description{
Get a node's partition element number.
}
