% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetNumberOfPartitions}
\alias{GetNumberOfPartitions}
\title{Get number of partitions.}
\usage{
GetNumberOfPartitions(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Calculate the number of partitions for a given labelled partition. This is
'm' in Kuipers & Moffa (2015).
}
