% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetParentCombinations}
\alias{GetParentCombinations}
\title{Get parent combinations for a given node.}
\usage{
GetParentCombinations(partitioned_nodes, node, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{node}{Node name.}

\item{scorer}{A scorer object.}
}
\value{
List of parent combinations.
}
\description{
Get parent combinations for a given node.
}
