% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{GetParentsKey}
\alias{GetParentsKey}
\title{Get parents key.}
\usage{
GetParentsKey(parents, nodes)
}
\arguments{
\item{parents}{A character vector of the parent nodes.}

\item{nodes}{A character vector for all nodes.}
}
\description{
TODO: The in function is quite slow. May need to make this a for loop in C++.
}
