% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{GetPartiallyIncrementalEdges}
\alias{GetPartiallyIncrementalEdges}
\title{Get partially incremental scoring edges.}
\usage{
GetPartiallyIncrementalEdges(scorer, cutoff = 0)
}
\arguments{
\item{scorer}{A scorer object.}

\item{cutoff}{A cutoff value for the blacklist. Less than this value is
blacklisted.}
}
\description{
Get the positive incremental scoring edges after conditioning
on all other variables.
}
