% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetPartitionedNodesFromAdjacencyMatrix}
\alias{GetPartitionedNodesFromAdjacencyMatrix}
\title{Derived a labelled partition.}
\usage{
GetPartitionedNodesFromAdjacencyMatrix(adjacency)
}
\description{
This partitions nodes into levels of outpoints as explained in Section 4.1 of
Kuipers & Moffa 2015. This takes an adjacency matrix and returns a data.frame
of (partition, node) pairs
}
