% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetPartitionNodes}
\alias{GetPartitionNodes}
\title{Get nodes in a partition element.}
\usage{
GetPartitionNodes(partitioned_nodes, elements)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{elements}{An integer or vector of integers for the partition element number.}
}
\description{
Get nodes in a partition element.
}
