% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetRestrictedNodes}
\alias{GetRestrictedNodes}
\title{Get nodes that have restricted parents.}
\usage{
GetRestrictedNodes(list)
}
\arguments{
\item{list}{A black or white list.}
}
\description{
Get nodes that have restricted parents.
}
