% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetRestrictedParents}
\alias{GetRestrictedParents}
\title{Get black or white listed parents.}
\usage{
GetRestrictedParents(node, listed = NULL)
}
\arguments{
\item{node}{The name of the node to get white or black listed parents.}

\item{listed}{A black or white list.}
}
\description{
Get black or white listed parents.
}
