% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{InitCoupledPartition}
\alias{InitCoupledPartition}
\title{Initialise partition state for SampleChains.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
InitCoupledPartition(
  nodes,
  scorer,
  init_state = NULL,
  n_coupled_chains = 4,
  n_parallel_chains = 2
)
}
\arguments{
\item{nodes}{A character vector of node names.}

\item{scorer}{A scorer object.}

\item{init_state}{Coupled partition state. Default is NULL.}

\item{n_coupled_chains}{Number of coupled chains. Default is 4.}

\item{n_parallel_chains}{Number of parallel chains to be run in SampleChains.
Default is 2.}
}
\description{
Initialise partition state for SampleChains.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
scorer <- CreateScorer(
  scorer = BNLearnScorer,
  data = bnlearn::learning.test
  )
nodes <- names(bnlearn::learning.test)

InitCoupledPartition(nodes, scorer)

}
