% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{InitPartition}
\alias{InitPartition}
\title{Initialise states for SampleChains.
Initialise partition state for SampleChains.}
\usage{
InitPartition(nodes, scorer, init_state = NULL, n_parallel_chains = 2)
}
\arguments{
\item{nodes}{A character vector of node names.}

\item{scorer}{A scorer object.}

\item{init_state}{A data.frame representing a partition. Default is NULL.}

\item{n_parallel_chains}{Number of parallel chains to be run in SampleChains.
Default is 2.}
}
\description{
Initialise states for SampleChains.
Initialise partition state for SampleChains.
}
\examples{
data <- bnlearn::learning.test

scorer <- CreateScorer(
  scorer = BNLearnScorer,
  data = data
  )
InitPartition(colnames(data), scorer)

}
