% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{LogSumExp}
\alias{LogSumExp}
\title{Log-Sum-Exponential calculation using the trick that limits underflow issues.}
\usage{
LogSumExp(x)
}
\arguments{
\item{x}{A vector of numeric.}
}
\value{
Log-Sum-Exponential (LSE) of x.
}
\description{
Log-Sum-Exponential calculation using the trick that limits underflow issues.
}
