% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{OrderPartitionedNodes}
\alias{OrderPartitionedNodes}
\title{Order partitioned nodes.}
\usage{
OrderPartitionedNodes(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\value{
Labelled partitioned in descending partition element order.
}
\description{
Order partitioned nodes.
}
