% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_steps.R
\name{PartitionMCMC}
\alias{PartitionMCMC}
\title{A transition object.
A one-step implementation of partition MCMC. This acts as a constructor.}
\usage{
PartitionMCMC(proposal = DefaultProposal(), verbose = TRUE)
}
\arguments{
\item{proposal}{Proposal function. Default is the DefaultProposal.}

\item{verbose}{Flag to pass mcmc information.}

\item{current_state}{A labelled partition.}

\item{scorer}{Scorer object.}
}
\value{
A new state.
}
\description{
A transition object.
A one-step implementation of partition MCMC. This acts as a constructor.
}
\examples{
transition <- PartitionMCMC(proposal = ProposePartitionSplitJoin)
transition(current_state, scorer_1)

}
