% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_split_join.R
\name{PartitionSplitJoin}
\alias{PartitionSplitJoin}
\title{Partition split or join constructor.}
\usage{
PartitionSplitJoin(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Partition split or join constructor.
}
