% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{PlotConcordance}
\alias{PlotConcordance}
\title{Concordance plot as outl.}
\usage{
PlotConcordance(x, ...)
}
\arguments{
\item{x}{A list of edge probabilities, a chains object, or a collections object.}

\item{...}{Additional parameter to send to the appropriate method. This includes
'highlight' (defauled to 0.3) which sets the cutoff difference that is used to
highlight the points, and the probability edge estimation 'method' for a
dagmc_collections object.}
}
\description{
This is a thin wrapper to pass dagmc objects to the BiDAG concordance
plot code.
}
\examples{
data <- bnlearn::learning.test
dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)
scorer <- CreateScorer(scorer = BNLearnScorer, data = data)

results <- SampleChains(300, partitioned_nodes, PartitionMCMC(), scorer)
dags <- PartitiontoDAG(results, scorer)
p_edge <- CalculateEdgeProbabilities(dags)
g <- PlotConcordance(p_edge)

}
