% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_node_move.R
\name{ProposeNodeMove}
\alias{ProposeNodeMove}
\title{Propose individual node movement.}
\usage{
ProposeNodeMove(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
This proposes that a single node selected uniformly can either:
\enumerate{
\item Move to any current partition.
\item Move to any gap between or at the ends of the partitions.
}

Any of these moves are possible and are selected uniformly with two
exceptions:
\enumerate{
\item The selected node cannot move into adjacent gaps if it originated from
a single node partition.
\item The selected node cannot move to the immediately higher gap if it
originated from a two node partition.
}
}
\examples{
dag <- UniformlySampleDAG(c('A', 'B', 'C', 'D', 'E', 'F'))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)
ProposeNodeMove(partitioned_nodes)

}
