% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_split_join.R
\name{ProposePartitionSplitJoin}
\alias{ProposePartitionSplitJoin}
\title{Propose a split or join of two partitions.}
\usage{
ProposePartitionSplitJoin(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}
}
\value{
A proposed labelled partition.
}
\description{
This is the 'Basic Move' (i.e. algorithm 1) in Kuipers & Moffa (2015). There
is a caveat in that the split proposal for a partition with one element is
ambiguous, as a split for such a partition element results in a stay still
proposal. Such a proposal has been removed.
}
\examples{
dag <- UniformlySampleDAG(c('A', 'B', 'C', 'D', 'E', 'F'))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)
ProposePartitionSplitJoin(partitioned_nodes)

}
