% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_swap_node.R
\name{ProposeSwapNode}
\alias{ProposeSwapNode}
\title{Propose that two nodes swap partition elements.}
\usage{
ProposeSwapNode(partitioned_nodes, verbose = FALSE)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}

\item{verbose}{A flag to indicate that you are staying still.}
}
\value{
A proposed labelled partition.
}
\description{
Propose that two nodes swap partition elements.
}
\examples{
ProposeStayStill(partitioned_nodes, verbose = TRUE)

}
