% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{SampleChain}
\alias{SampleChain}
\title{Sample chain.}
\usage{
SampleChain(n_results, init_state, transition, scorer, n_thin = 1)
}
\arguments{
\item{n_results}{Number of saved states.}

\item{init_state}{An initial state that can be passed to transition.}

\item{transition}{A transition function.}

\item{scorer}{A scorer object.}

\item{n_thin}{Number of steps between saved states.}
}
\value{
List of results.
}
\description{
Sample chain.
}
\examples{
results <- SampleChain(100, partitioned_nodes, transition = PartitionMCMC(scorer = scorer_1))

}
