% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{SampleChainDAGs}
\alias{SampleChainDAGs}
\title{Sampled DAG from chains.}
\usage{
SampleChainDAGs(chains, scorer)
}
\arguments{
\item{chains}{MCMC chains.}

\item{scorer}{Scorer object.}
}
\value{
Chains with sample dags and their corresponding score.
}
\description{
Sampled DAG from chains.
}
