% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dag_from_labelled_partition.R
\name{SampleDAGFromLabelledPartition}
\alias{SampleDAGFromLabelledPartition}
\title{Sample a DAG from a labelled partition.}
\usage{
SampleDAGFromLabelledPartition(partitioned_nodes, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{scorer}{Scorer object.}
}
\value{
A list with elements:
\itemize{
\item state Adjacency matrix.
\item log_score Score of the sampled DAG.
}
}
\description{
Sample a single DAG with  from the set of DAGs that is consistent with a labelled
partition. That is, given a set of DAGs \eqn{\mathcal{G}} that is consistent
with a labelled partition \eqn{\Lambda} then it will sample a given DAG
\eqn{G} where \eqn{G \in \mathcal{G}} with probability
\eqn{p(G|D, \Lambda) = p(G)p(D | G) / \sum_{G \in \mathcal{G}} p(G)p(D | G)}.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer <- CreateScorer(data = data)

SampleDAGFromLabelledPartition(partitioned_nodes, scorer)

}
