% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_edge_probabilities.R
\name{SampleEdgeProbabilities}
\alias{SampleEdgeProbabilities}
\title{Sample edge probabilities}
\usage{
SampleEdgeProbabilities(x)
}
\arguments{
\item{x}{A chain(s) or collection object where states are DAGs.}
}
\value{
p_edge A posterior sample for the marginalised edge probabilities.
}
\description{
Sample edge probabilities
}
\examples{
data <- bnlearn::learning.test

scorer <- CreateScorer(
  scorer = BNLearnScorer, 
  data = data
  )

init_state <- InitPartition(colnames(data), scorer)
results <- SampleChains(10, init_state, PartitionMCMC(), scorer)
dag_chains <- PartitiontoDAG(results, scorer)

pedge_sample <- SampleEdgeProbabilities(dag_chains)

}
