% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreDiff}
\alias{ScoreDiff}
\title{Calculate the difference in log scores between two labelled partitions.}
\usage{
ScoreDiff(old_partitioned_nodes, new_partitioned_nodes, scorer)
}
\arguments{
\item{old_partitioned_nodes}{A labelled partition.}

\item{new_partitioned_nodes}{A labelled partition.}

\item{scorer}{A scorer object.}
}
\value{
changed_nodes A vector of changed nodes.
}
\description{
Calculate the difference in log scores between two labelled partitions.
}
\examples{
ScoreDiff(partitioned_nodes, new_partitioned_nodes, scorer = scorer_1)
ScoreDiff(partitioned_nodes, new_partitioned_nodes, scorer = scorer_2)

}
