% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreLabelledPartition}
\alias{ScoreLabelledPartition}
\title{Score labelled partition}
\usage{
ScoreLabelledPartition(partitioned_nodes, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{scorer}{Scorer object.}
}
\value{
Log of the node score.
}
\description{
Score labelled partition
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(names(data))
partitioned_nodes <- DAGtoPartition(dag)

scorer <- CreateScorer(
  scorer = BNLearnScorer, 
  data = data
  )

ScoreLabelledPartition(partitioned_nodes, scorer)

}
