% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreTableNode}
\alias{ScoreTableNode}
\title{Calculate score tables for (node, parents) combinations.}
\usage{
ScoreTableNode(partitioned_nodes, node, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{node}{Name of node.}

\item{scorer}{Scorer object.}
}
\value{
List of log_scores for each combination in parent_combinations.
}
\description{
Calculate score tables for (node, parents) combinations.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(names(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )
  
ScoreTableNode(partitioned_nodes, 'A', scorer)

}
