% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_stay.R
\name{StayStill}
\alias{StayStill}
\title{StayStill proposal.}
\usage{
StayStill(partitioned_nodes, verbose = FALSE)
}
\arguments{
\item{verbose}{A flag to indicate that you are staying still.}
}
\description{
StayStill proposal.
}
