% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{[.cia_chain}
\alias{[.cia_chain}
\title{Index a cia_chain object}
\usage{
\method{[}{cia_chain}(x = list(), i, ...)
}
\arguments{
\item{x}{A cia_chain object.}

\item{i}{An index.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
A cia_chain.
}
\description{
Index a cia_chain object
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- DAGtoPartition(dag)

scorer <- CreateScorer(
  scorer = BNLearnScorer, 
  data = data
  )

results <- SampleChains(10, partitioned_nodes, PartitionMCMC(), scorer)
results[[1]][5]

}
