% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{[.cia_chains}
\alias{[.cia_chains}
\title{Index a cia_chains object}
\usage{
\method{[}{cia_chains}(x = list(), i, ...)
}
\arguments{
\item{x}{A cia_chain object.}

\item{i}{An index to get the cia_chain iterations.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A cia_chains object.
}
\description{
Index a cia_chains object
}
