% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_chains.R
\name{[.cia_post_chain}
\alias{[.cia_post_chain}
\title{Indexing with respect to iterations.}
\usage{
\method{[}{cia_post_chain}(x = list(), i, ...)
}
\arguments{
\item{x}{A cia_post_chain object.}

\item{i}{An index.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A cia_post_chain.
}
\description{
Indexing with respect to iterations.
}
