% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{[.dagmc_chain}
\alias{[.dagmc_chain}
\title{Indexing with respect to iterations.}
\usage{
\method{[}{dagmc_chain}(x = list(), i, ...)
}
\arguments{
\item{x}{A dagmc_chain object.}

\item{i}{An index.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A dagmc_chain.
}
\description{
Indexing with respect to iterations.
}
