% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{[.dagmc_chains}
\alias{[.dagmc_chains}
\title{Index a dagmc_chains object with respect to iterations.}
\usage{
\method{[}{dagmc_chains}(x = list(), i, ...)
}
\arguments{
\item{x}{A dagmc_chain object.}

\item{i}{An index to get the dagmc_chain iterations.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A dagmc_chains object.
}
\description{
Index a dagmc_chains object with respect to iterations.
}
