% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_chains.R
\name{[[.cia_post_chains}
\alias{[[.cia_post_chains}
\title{Index a cia_post_chains object.}
\usage{
\method{[[}{cia_post_chains}(x, i, ...)
}
\arguments{
\item{x}{A cia_post_chains object.}

\item{i}{An index to get the cia_post_chain.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A cia_post_chains object.
}
\description{
Index a cia_post_chains object.
}
