% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{[[.dagmc_chains}
\alias{[[.dagmc_chains}
\title{Index a dagmc_chains object.}
\usage{
\method{[[}{dagmc_chains}(x, i, ...)
}
\arguments{
\item{x}{A dagmc_chains object.}

\item{i}{An index to get the dagmc_chain.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A dagmc_chains object.
}
\description{
Index a dagmc_chains object.
}
