% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_swap_adjacent.R
\name{SwapAdjacentNode}
\alias{SwapAdjacentNode}
\title{Swap nodes from adjacent partition elements proposal.}
\usage{
SwapAdjacentNode(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}
}
\description{
Swap nodes from adjacent partition elements proposal.
}
