% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_steps.R
\name{TemperedPartitionMCMC}
\alias{TemperedPartitionMCMC}
\title{One step implementation of the tempered partition MCMC. This acts as a constructor.}
\usage{
TemperedPartitionMCMC(proposal = NULL, temperature = 1, verbose = TRUE)
}
\arguments{
\item{proposal}{Proposal function. Default is the DefaultProposal.}

\item{temperature}{Numeric value representing the temperature to raise the
score to.}

\item{verbose}{Flag to pass MCMC information.}
}
\value{
Function that takes the current state and scorer that outputs a new
state.
}
\description{
This is a constructor for a single Tempered Partition MCMC step. The function
constructs an environment with the proposal, inverse temperature, and verbose
flag. It then returns a function which takes the current_state and a scorer
object. This only allows the scores to be raised to a constant temperature
for every step.
}
\examples{
dag <- UniformlySampleDAG(c('A', 'B', 'C', 'D', 'E', 'F'))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer <- list(
  scorer = BNLearnScorer,
  parameters = list(data = bnlearn::learning.test)
  )

current_state <- list(
  state = partitioned_nodes,
  log_score = ScoreLabelledPartition(partitioned_nodes, scorer)
  )

tpmcmc <- TemperedPartitionMCMC(proposal = NodeMove, temperature = 1.0)
tpmcmc(current_state, scorer)

}
