% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_bnlearn.R
\name{toBNLearn}
\alias{toBNLearn}
\title{Convert to bnlearn object.}
\usage{
toBNLearn(x)
}
\arguments{
\item{x}{An object that represents a DAG.}
}
\value{
bn_obj A bn object.
}
\description{
Convert to bnlearn object.
}
\examples{
adj <- UniformlySampleDAG(c('A', 'B', 'C'))
toBNLearn(adj)

}
