% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_grain.R
\name{togRain}
\alias{togRain}
\title{Convert to a gRain object.}
\usage{
togRain(x, ...)
}
\arguments{
\item{x}{An adjacency matrix or igraph object.}

\item{...}{extra parameters to gRain compile.}
}
\value{
A gRain object.
}
\description{
Convert to a gRain object.
}
\examples{
dag <- bnlearn::model2network("[A][C][F][B|A][D|A:C][E|B:F]")
gRain_obj <- togRain(x = dag |> toMatrix(), data = bnlearn::learning.test)

}
