% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{toMatrix}
\alias{toMatrix}
\title{Convert bnlearn network object to an adjacency matrix.}
\usage{
toMatrix(network)
}
\arguments{
\item{network}{A network object from bnlearn or igraph.}
}
\value{
An adjacency matrix representation of network.
}
\description{
Convert bnlearn network object to an adjacency matrix.
}
\examples{
toMatrix(bnlearn::empty.graph(LETTERS[1:6]))

}
