% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dags.R
\name{UniformlySampleDAG}
\alias{UniformlySampleDAG}
\title{Uniformly sample DAG}
\usage{
UniformlySampleDAG(nodes)
}
\arguments{
\item{nodes}{A vector of node names.}
}
\value{
Adjacency matrix with elements designated as (parent, child).
}
\description{
Uniformly sample DAG
}
\examples{
UniformlySampleDAG(LETTERS[1:3])

}
