% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{CalculateEdgeProbabilities}
\alias{CalculateEdgeProbabilities}
\title{Calculate marginalised edge probabilities.}
\usage{
CalculateEdgeProbabilities(collection)
}
\arguments{
\item{collection}{A collection of objects. See CollectUniqueObjects.}
}
\value{
p_edge An adjacency matrix representing the edge probabilities.
}
\description{
Calculate the probability of a given edge (\eqn{E}) given the data which
is given by, \deqn{p(E|D) = \sum_\mathcal{G} p(E|G)p(G|D)}).
}
