% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_probabilities.R
\name{CalculateEdgeProbabilities}
\alias{CalculateEdgeProbabilities}
\title{Calculate pairwise edge probabilities marginalised over the graph structure.}
\usage{
CalculateEdgeProbabilities(x, ...)
}
\arguments{
\item{x}{A chain(s) or collection object where states are DAGs.}

\item{...}{Extra parameters sent to the methods. For a dag collection you can
choose to use method='sampled' for MCMC sampled frequency (which is our
recommended method) or method='score' which uses the normalised scores.}
}
\value{
p_edge A matrix representing the edge probabilities.
}
\description{
Calculate pairwise edge probabilities. For a given edge from node
\eqn{A \rightarrow B}, the posterior probability of that edge given the data
is given by marginalising out the graph structure \eqn{G}, such that
\deqn{p(E|D) = \sum_{G \in \mathcal{G}} p(E|G)p(G|D).}
}
