% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_mean.R
\name{CalculateFeatureMean}
\alias{CalculateFeatureMean}
\title{Calculate arithmetic mean for a DAG feature marginalised over DAGs.}
\usage{
CalculateFeatureMean(x, p_feature, ...)
}
\arguments{
\item{x}{A chain(s) or collection object.}

\item{p_feature}{A function that takes an adjacency matrix or collection object
and returns a scalar corresponding to \eqn{p(f|G)}. The function must be of the
form p_feature(dag).}

\item{...}{Extra parameters sent to the methods. For a dag collection you can
choose to use method='sampled' for MCMC sampled frequency (which is our
recommended method) or method='score' which uses the normalised scores.}
}
\value{
p_post_feature A numeric value representing the posterior probability
of the feature.
}
\description{
Calculate the posterior expected value for a feature (\eqn{f}, e.g.,
existence of an edge) whereby \deqn{E(f|D) = \sum_{G \in \mathcal{G}} p(f|G) p(G|D).} This can be
useful for calculating point estimates of quantities of interests, such as
the probability that an edge exists.
}
