% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_probabilities.R
\name{CalculateFeatureProbability}
\alias{CalculateFeatureProbability}
\title{Collect feature probability marginalised over states.}
\usage{
CalculateFeatureProbability(x, p_feature, ...)
}
\arguments{
\item{x}{A chain(s) or collection object.}

\item{p_feature}{A function that takes an adjacency matrix and collection object
and returns a numeric value equal to p(f|G). Therefore, it must be of the
form p_feature(dag).}

\item{...}{Extra parameters sent to the methods. For a dag collection you can
choose to use method='sampled' for MCMC sampled frequency (which is our
recommended method) or method='score' which uses the normalised scores.}
}
\value{
p_post_feature A numeric value representing the posterior probability
of the feature.
}
\description{
Calculate the feature (\eqn{f}) probability whereby
\eqn{p(f|D) = \sum_\mathcal{G \in G} p(G|D)p(f|G)}.
}
