% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_package.R
\docType{package}
\name{cia-package}
\alias{cia}
\alias{cia-package}
\title{cia: Causal inference assistant (CIA)}
\description{
A library for performing causal inference within the structural causal modelling framework. Structure learning is performed using partition MCMC (Kuipers & Moffa, 2017) and several additional functions have been added to help with causal inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spaceodyssey.github.io/cia/}
}

}
\author{
\strong{Maintainer}: Mathew Varidel \email{mathew.varidel@sydney.edu.au} (\href{https://orcid.org/0000-0002-1648-8317}{ORCID})

}
\keyword{internal}
