% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmc_package.R
\docType{package}
\name{dagmc-package}
\alias{dagmc}
\alias{dagmc-package}
\title{dagmc: MCMC sampler for Directed Acyclic Graphs}
\description{
An implementation of the partition MCMC algorithm for Directed Acyclic Graphs. The default scoring function uses bnlearn, which means that you have access to all of their scoring options and functionality. References: J. Kuipers and G. Moffa (2017) \doi{10.1080/01621459.2015.1133426}, M. Scutari (2010) \doi{10.18637/jss.v035.i03}.
}
\author{
\strong{Maintainer}: Mathew Varidel \email{mathew.varidel@sydney.edu.au} (\href{https://orcid.org/0000-0002-1648-8317}{ORCID})

}
\keyword{internal}
