% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagmc_package.R
\docType{package}
\name{dagmc-package}
\alias{dagmc}
\alias{dagmc-package}
\title{dagmc: MCMC sampler for Directed Acyclic Graphs}
\description{
An implementation of the partition MCMC algorithm for Directed Acyclic Graphs. The default implementation should be consistent with BiDAG with some extra functionality allowing you to create your own proposals. The default scoring function uses bnlearn, which means that you have access to all of their scoring options and functionality. You can also build your own scorer if you wish to do so.
}
\author{
\strong{Maintainer}: Mathew Varidel \email{mathew.varidel@sydney.edu.au} (\href{https://orcid.org/0000-0002-1648-8317}{ORCID})

}
\keyword{internal}
