% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_to_cpdag.R
\name{DAGtoCPDAG}
\alias{DAGtoCPDAG}
\title{Convert DAG to CPDAG}
\usage{
DAGtoCPDAG(x)
}
\arguments{
\item{x}{A matrix, cia_chain, or cia_chains object. When it is a chain(s)
object the state must be an adjacency matrix.}
}
\value{
x Returns same object type converted to a CPDAG.
}
\description{
Converts a directed acyclic graph (DAG) into it's equivalence class
corresponding to a completed partially directed acyclic graph (CPDAG).
}
\examples{
dag <- UniformlySampleDAG(LETTERS[1:3])
DAGtoCPDAG(dag)

}
