% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_proposal.R
\name{DefaultProposal}
\alias{DefaultProposal}
\title{Default proposal}
\usage{
DefaultProposal(
  p_split_join = 0.4,
  p_swap_node = 0.3,
  p_node_move = 0.29,
  verbose = FALSE
)
}
\description{
Default proposal
}
