% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_proposal.R
\name{DefaultProposal}
\alias{DefaultProposal}
\title{Default proposal constructor.}
\usage{
DefaultProposal(p = c(0.33, 0.33, 0.33, 0.01), verbose = TRUE)
}
\arguments{
\item{p}{Probability for each proposal in the order (split_join, node_move,
swap_node, stay_still).}

\item{verbose}{Boolean flag to record proposal used.}
}
\description{
Default proposal constructor.
}
