% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{FindChangedNodes}
\alias{FindChangedNodes}
\title{Find nodes with changed parent combinations between different labelled
partitions.}
\usage{
FindChangedNodes(old_partitioned_nodes, new_partitioned_nodes, scorer)
}
\arguments{
\item{old_partitioned_nodes}{Labelled partition.}

\item{new_partitioned_nodes}{Labelled partition.}

\item{scorer}{Scorer object.}
}
\value{
Vector of changed nodes.
}
\description{
TODO: This is quite slow. From the proposal we should be able to determine
the nodes that need to be rescored rather than finding them using this
function.
}
\examples{
scorer = CreateScorer()

old_dag <- UniformlySampleDAG(LETTERS[1:5])
old_partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(old_dag)

new_dag <- UniformlySampleDAG(LETTERS[1:5])
new_partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(new_dag)

changed_nodes <- FindChangedNodes(old_partitioned_nodes, new_partitioned_nodes, scorer)

}
