% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{GetIncrementalScoringEdges}
\alias{GetIncrementalScoringEdges}
\title{Get the score of the empty DAG}
\usage{
GetIncrementalScoringEdges(scorer, cutoff = 0)
}
\arguments{
\item{scorer}{A scorer object.}

\item{cutoff}{A score cutoff.}
}
\value{
A Boolean matrix of (parent, child) pairs for blacklisting..
}
\description{
Get the score of the empty DAG
}
