% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{GetLowestPairwiseScoringEdges}
\alias{GetLowestPairwiseScoringEdges}
\title{Preprocessing for blacklisting.
Get the lowest pairwise scoring edges.}
\usage{
GetLowestPairwiseScoringEdges(scorer, n_retain)
}
\arguments{
\item{scorer}{A scorer object.}

\item{n_retain}{An integer representing the number of edges to retain.}
}
\value{
A boolean matrix of (parent, child) pairs for blacklisting.
}
\description{
Get the lowest c pairwise scoring edges represented as a blacklist matrix.
This blacklisting procedure is motivated by Koller & Friedman (2003).
}
\references{
\enumerate{
\item Koller D, Friedman N. Being Bayesian about network structure. A Bayesian
approach to structure discovery in Bayesian networks. Mach Learn.
2003;50(1):95–125.
}
}
