% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{GetMAP}
\alias{GetMAP}
\title{Get the maximum a posteriori state.}
\usage{
GetMAP(x)
}
\arguments{
\item{x}{A collection of unique objects or chains object.}
}
\value{
maps A list with the adjacency matrix for the map and it's posterior
probability. It is possible for it to return multiple DAGs. The list has
elements;
\itemize{
\item state: List of MAP DAGs.
\item log_p: Numeric vector with the log posterior probability for each state.
}
}
\description{
Get the maximum a posteriori state.
}
