% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetNodePartition}
\alias{GetNodePartition}
\title{Get a node's partition element number.}
\usage{
GetNodePartition(node, partitioned_nodes)
}
\arguments{
\item{node}{A node name.}

\item{partitioned_nodes}{A labelled partition.}
}
\value{
node_partition A node's partition element number.
}
\description{
Get a node's partition element number.
}
