% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetOrderedPartition}
\alias{GetOrderedPartition}
\title{Get ordered labelled partition.}
\usage{
GetOrderedPartition(partitioned_nodes)
}
\description{
Calculate the ordered partition. Labelled as lamba in Kuipers & Moffa (2015).
}
