% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{GetParentCombinations}
\alias{GetParentCombinations}
\title{Get all parent combinations of nodes.}
\usage{
GetParentCombinations(parents)
}
\arguments{
\item{parents}{A vector of potential parent names.}
}
\value{
parent_combinations A list of parent combinations.
}
\description{
Get all parent combinations of nodes.
}
