% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_steps.R
\name{PartitionMCMC}
\alias{PartitionMCMC}
\title{A transition object.
A one-step implementation of partition MCMC.}
\usage{
PartitionMCMC(current_state, proposal, scorer)
}
\arguments{
\item{current_state}{A labelled partition.}

\item{proposal}{A proposal function.}

\item{scorer}{A scorer object.}
}
\value{
A new state.
}
\description{
A transition object.
A one-step implementation of partition MCMC.
}
\examples{
PartitionMCMC(partitioned_nodes, proposal = ProposePartitionSplitJoin, scorer = scorer_1)
PartitionMCMC(partitioned_nodes, proposal = ProposePartitionSplitJoin, scorer = scorer_2)

}
