% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_to_dag.R
\name{PartitiontoDAG}
\alias{PartitiontoDAG}
\title{Sample DAGs from labelled partitions.}
\usage{
PartitiontoDAG(partitions, scorer)
}
\arguments{
\item{partitions}{A dagms_chains, dagmc_chain, or matrix.}

\item{scorer}{A scorer object.}
}
\description{
Sample DAGs from labelled partitions.
}
