% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_to_dag.R
\name{PartitiontoDAG}
\alias{PartitiontoDAG}
\title{Sample DAGs from labelled partitions.}
\usage{
PartitiontoDAG(partitions, scorer)
}
\arguments{
\item{partitions}{A cia_chain(s) object or adjacency matrix.}

\item{scorer}{A scorer object.}
}
\description{
Sample DAGs from labelled partitions.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partition <- DAGtoPartition(dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )

# Used to sample from a single partition.  
PartitiontoDAG(partition, scorer)

# Used to convert a chain of partitions to DAGs.
results <- SampleChains(3, partition, PartitionMCMC(), scorer)
PartitiontoDAG(results, scorer)

}
