% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummean.R
\name{PlotCumulativeMeanTrace}
\alias{PlotCumulativeMeanTrace}
\title{Plot cumulative mean trace plot.}
\usage{
PlotCumulativeMeanTrace(
  x,
  ncol = NULL,
  nrow = NULL,
  scales = "fixed",
  dir = "v"
)
}
\arguments{
\item{x}{A posterior predictive sample object.}

\item{ncol}{Number of columns.}

\item{nrow}{Number of rows.}

\item{scales}{Whether the scales should the fixed ('fixed', the default),
free ('free') or free in one dimension ('free_x', 'free_y')?}

\item{dir}{Direction to fill facets. Either 'h' for horizontal or 'v' for
vertical.}
}
\description{
Plot cumulative mean trace plot.
}
