% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{PostProcessChains}
\alias{PostProcessChains}
\title{Analysis of chains.
Equilibrium states.}
\usage{
PostProcessChains(chains, n_burnin = 0, n_thin = 1)
}
\arguments{
\item{chains}{MCMC chains.}

\item{n_burnin}{Number of steps to remove at the start as a burnin. Default is 0.}

\item{n_thin}{Number of steps between retained states. Default is 1.}
}
\description{
This allows you to remove a burnin and thin the chains after processing.
}
