% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_node_move.R
\name{ProposeNodeMove}
\alias{ProposeNodeMove}
\title{Propose individual node movement.}
\usage{
ProposeNodeMove(partitioned_nodes, verbose = FALSE)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}

\item{verbose}{A flag to indicate that a single node is being moved.}
}
\description{
This proposes that a single node selected uniformly can either:
\enumerate{
\item Move to any current partition element
\item Move to any gap between or at the ends of the elements.
}

Any of these moves are possible and are selected uniformly. Note that
this differs slightly from KP15 who remove the possibility of a 2-node
partition element . This may slow the convergence on a per step basis but is
easier and more concise to implement.
}
\examples{
ProposeNodeMovement(partitioned_nodes, verbose = TRUE)

}
