% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{ProposePartitionSplitJoin}
\alias{ProposePartitionSplitJoin}
\title{Propose a split or join of two partitions.}
\usage{
ProposePartitionSplitJoin(partitioned_nodes, verbose = FALSE)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}

\item{verbose}{A flag to indicate whether to inform the user that the move
is a split or join.}
}
\value{
A proposed labelled partition.
}
\description{
This is the 'Basic Move' (i.e. algorithm 1) in Kuipers & Moffa (2015).
}
\examples{
ProposePartitionSplitJoin(partitioned_nodes, verbose = TRUE)

}
