% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_stay.R
\name{ProposeStayStill}
\alias{ProposeStayStill}
\title{Propose that the partition stays still.}
\usage{
ProposeStayStill(partitioned_nodes)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}
}
\value{
A proposed labelled partition.
}
\description{
Propose that the partition stays still.
}
\examples{
ProposeStayStill(partitioned_nodes)

}
