% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_stay.R
\name{ProposeStayStill}
\alias{ProposeStayStill}
\title{Propose that the partition stays still.}
\usage{
ProposeStayStill(partitioned_nodes, verbose = FALSE)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}

\item{verbose}{A flag to indicate that you are staying still.}
}
\value{
A proposed labelled partition.
}
\description{
Propose that the partition stays still.
}
\examples{
ProposeStayStill(partitioned_nodes, verbose = TRUE)

}
