% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{SampleChain}
\alias{SampleChain}
\title{Sample chain.}
\usage{
SampleChain(n_results, init_state, transition, proposal, scorer, n_thin = 1)
}
\arguments{
\item{n_results}{Number of saved states.}

\item{init_state}{An initial state that can be passed to transition.}

\item{transition}{A transition function.}

\item{scorer}{A scorer object.}

\item{n_thin}{Number of steps between saved states.}
}
\value{
A list of states and log_score.
}
\description{
Sample chain.
}
\examples{
trace <- SampleChain(100, partitioned_nodes, transition = PartitionMCMC, 
proposal = ProposePartitionSplitJoin, scorer = scorer_1, 
n_thin = 1)

}
