% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{SampleChain}
\alias{SampleChain}
\title{Sample a single chain.}
\usage{
SampleChain(n_results, init_state, transition, scorer, n_thin = 1)
}
\arguments{
\item{n_results}{Number of saved states.}

\item{init_state}{An initial state that can be passed to transition.}

\item{transition}{A transition function.}

\item{scorer}{A scorer object.}

\item{n_thin}{Number of steps between saved states.}
}
\value{
chain A dagmc_chain object.
}
\description{
Sample a single chain.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer_1 <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )

results <- SampleChain(10, partitioned_nodes, PartitionMCMC(), scorer_1)

}
