% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dag_from_labelled_partition.R
\name{SampleDAGFromLabelledPartition}
\alias{SampleDAGFromLabelledPartition}
\title{Sample a DAG from a labelled partition.}
\usage{
SampleDAGFromLabelledPartition(partitioned_nodes, scorer)
}
\arguments{
\item{partitioned_nodes}{A labelled partition.}

\item{scorer}{A scorer object.}
}
\value{
A data.frame of (node, partition element) pairs.
}
\description{
Sample a DAG from a labelled partition.
}
\examples{
SampleDAGFromLabelledPartition(partitioned_nodes, scorer = scorer_1)

}
