% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dag_from_labelled_partition.R
\name{SampleDAGFromLabelledPartition}
\alias{SampleDAGFromLabelledPartition}
\title{Sample a DAG from a labelled partition.}
\usage{
SampleDAGFromLabelledPartition(partitioned_nodes, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{scorer}{Scorer object.}
}
\value{
A matrix with elements of (parent, child) cells with 1 representing
and edge and 0 otherwise.
}
\description{
Sample a DAG from a labelled partition.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer <- CreateScorer(data = data)

SampleDAGFromLabelledPartition(partitioned_nodes, scorer = scorer)

}
