% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_edge_probabilities.R
\name{SampleEdgeProbabilities}
\alias{SampleEdgeProbabilities}
\title{Sample edge probabilities using chains object.}
\usage{
SampleEdgeProbabilities(x)
}
\arguments{
\item{x}{A chain(s) or collection object where states are DAGs.}
}
\value{
p_edge A posterior sample for the marginalised edge probabilities.
}
\description{
Sample edge probabilities using chains object.
}
