% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{SamplePosteriorPredictiveChains}
\alias{SamplePosteriorPredictiveChains}
\title{Draw from a posterior predictive distribution.}
\usage{
SamplePosteriorPredictiveChains(x, p_predict, ...)
}
\arguments{
\item{x}{A cia_chain(s) object.}

\item{p_predict}{A function that draws from the posterior predictive distribution
of interest given an adjacency matrix representing a DAG. The function must
be of the form p_predict(dag, ...) and return either a vector of numeric values.}

\item{...}{Parameters to be passed to p_predict.}
}
\description{
Sample a feature given a DAG (\eqn{p(f|G}) for each posterior sample.
}
