% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreDAG}
\alias{ScoreDAG}
\title{Score DAG.}
\usage{
ScoreDAG(dag, scorer)
}
\arguments{
\item{dag}{An adjacency matrix with (parent, child) entries. With 1 denoting
and edge and 0 otherwise.}
}
\value{
Log of DAG score.
}
\description{
Score DAG.
}
