% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreDiff}
\alias{ScoreDiff}
\title{Calculate the difference in log scores between two labelled partitions.}
\usage{
ScoreDiff(
  old_partitioned_nodes,
  new_partitioned_nodes,
  scorer,
  rescore_nodes = NULL
)
}
\arguments{
\item{old_partitioned_nodes}{A labelled partition.}

\item{new_partitioned_nodes}{A labelled partition.}

\item{scorer}{A scorer object.}

\item{rescore_nodes}{Default is NULL which will determine the}
}
\value{
Log of score difference between two labelled partitions.
}
\description{
Calculate the difference in log scores between two labelled partitions.
}
\examples{
data <- bnlearn::learning.test

old_dag <- UniformlySampleDAG(names(data))
old_partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(old_dag)

new_dag <- UniformlySampleDAG(names(data))
new_partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(new_dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )

ScoreDiff(old_partitioned_nodes, new_partitioned_nodes, scorer = scorer)

}
