% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreLabelledPartition}
\alias{ScoreLabelledPartition}
\title{Score labelled partition by adding the log scores for each node.}
\usage{
ScoreLabelledPartition(partitioned_nodes, scorer)
}
\arguments{
\item{scorer}{A scorer object.}

\item{node}{The node name.}

\item{potential_parents}{A vector of potential parents that need to be iterated over.}
}
\value{
log_node_score The log of the node score.
}
\description{
Score labelled partition by adding the log scores for each node.
}
\examples{
ScoreLabelledPartition(partitioned_nodes, scorer_1)
ScoreLabelledPartition(partitioned_nodes, scorer_2)

}
