% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreNode}
\alias{ScoreNode}
\title{Score node by marginalising over parent combinations.}
\usage{
ScoreNode(node, potential_parents, scorer)
}
\arguments{
\item{node}{The node name.}

\item{potential_parents}{A vector of potential parents that need to be iterated over.}

\item{scorer}{A scorer object.}
}
\value{
log_node_score The log of the node score.
}
\description{
Score node by marginalising over parent combinations.
}
\examples{
ScoreNode('A', c('B', 'C'), scorer_1)
ScoreNode('A', c('B', 'C'), scorer_2)

}
