% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreNode}
\alias{ScoreNode}
\title{Score node by marginalising over parent combinations.}
\usage{
ScoreNode(partitioned_nodes, node, scorer)
}
\arguments{
\item{partitioned_nodes}{Labelled partition.}

\item{node}{The node name.}

\item{scorer}{A scorer object.}
}
\value{
Log of the node score.
}
\description{
Score node by marginalising over parent combinations.
}
\examples{
data <- bnlearn::learning.test

dag <- UniformlySampleDAG(names(data))
partitioned_nodes <- GetPartitionedNodesFromAdjacencyMatrix(dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )
  
ScoreNode(partitioned_nodes, 'A', scorer)

}
