% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_funs.R
\name{ScoreTableNode}
\alias{ScoreTableNode}
\title{Calculate score tables for (node, parents) combinations.}
\usage{
ScoreTableNode(partitioned_nodes, node, scorer)
}
\arguments{
\item{node}{The node name.}

\item{scorer}{A scorer object.}

\item{potential_parents}{A vector of potential parents that need to be iterated over.}
}
\value{
List of log_scores for each combination in parent_combinations.
}
\description{
Calculate score tables for (node, parents) combinations.
}
\examples{
ScoreTableNode('A', c('B', 'C'), scorer_1)
ScoreTableNode('A', c('B', 'C'), scorer_2)
ScoreTableNode('A', c(), scorer_1)
ScoreTableNode('A', NULL, scorer_1)

}
