% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_chain.R
\name{[.cia_chain}
\alias{[.cia_chain}
\title{Indexing with respect to iterations.}
\usage{
\method{[}{cia_chain}(x = list(), i, ...)
}
\arguments{
\item{x}{A cia_chain object.}

\item{i}{An index.}

\item{...}{ellipsis for extra indexing parameters.}
}
\value{
chain A cia_chain.
}
\description{
Indexing with respect to iterations.
}
