% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_bnlearn.R
\name{toBNLearn}
\alias{toBNLearn}
\title{Convert adjacency matrix to bnlearn object.}
\usage{
toBNLearn(x)
}
\arguments{
\item{x}{Adjacency matrix.}
}
\value{
bn_obj A bn object.
}
\description{
Convert adjacency matrix to bnlearn object.
}
\examples{
adj <- UniformlySampleDAG(c('A', 'B', 'C'))
bn_obj <- toBNLearn(adj)

}
