% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_matrix.R
\name{toMatrix}
\alias{toMatrix}
\title{Convert to adjacency matrix.}
\usage{
toMatrix(network)
}
\arguments{
\item{network}{A bnlearn or igraph object.}
}
\value{
An adjacency matrix representation of network.
}
\description{
Convert a DAG object from other libraries to an adjacency matrix.
}
\examples{
toMatrix(bnlearn::empty.graph(LETTERS[1:6]))
toMatrix(igraph::sample_k_regular(10, 2))

}
